feat: refactor App component and create MainPage component for improved structure
